backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
supabase>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
stripe>=7.0.0
slowapi>=0.1.9
jsonschema>=4.20.0
bleach>=6.1.0
structlog>=23.2.0
pyjwt>=2.8.0
cryptography>=41.0.0

